requirements.txt
ebaysdk==2.2.0
fastapi==0.115.11
firebase_admin==6.7.0
protobuf==6.30.1
pydantic==2.10.6
python-dotenv==1.0.1
Requests==2.32.3
uvicorn==0.34.0
slowapi==0.1.9